cleaned up unused classes
