+++ dev/Client/src/components/App.js
@@ -17,10 +17,7 @@ import MyTransaction from './MyTransaction';
 import Register from './Register';
 import ManageStores from './ManageStores'
 import ManageStore from './ManageStore'
-import { ROUTE_MANAGE_EMPLOYEE } from '../routes';
-import ManageEmployee from './ManageEmployee';
 import ManageSystem from './system_manager/ManageSystem';
-import Checkout2 from './Checkout2';
 
 
 //import {BrowserRouter as Router, Route} from 'react-router-dom'
@@ -116,7 +113,7 @@ class App extends React.Component
           <Route
             path='/checkout' exact
             render={(props) => (
-              <Checkout2 {...props} getAppState={this.getAppState} setAppState={this.setAppState} />
+              <Checkout {...props} getAppState={this.getAppState} setAppState={this.setAppState} />
             )}
           />
           <Route
